Fix for Sendgrid SMTP-ID bug
In cases where the same recipient is defined twice, Sendgrid is sending
unique messages with the same recipient SMTP ID. This causes Gmail to
discard the extras, even if they have unique content (due to different
merge vars or etc.). This is a workaround for the issue until Sendgrid
decides what, if anything, they'd like to do on their end to resolve
this.